requirements.txt
absl-py==1.4.0
aiofiles==0.8.0
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
carvekit==4.1.0
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.1.3
efficientnet==1.0.0
fastapi==0.78.0
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.55.0
h11==0.14.0
h5py==3.8.0
idna==3.4
image-classifiers==1.0.0
imageio==2.28.1
keras==2.13.1rc0
Keras-Applications==1.0.8
lazy_loader==0.2
libclang==16.0.0
loguru==0.6.0
Markdown==3.4.3
MarkupSafe==2.1.2
networkx==3.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==23.1
Pillow==9.0.1
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.9.2
python-multipart==0.0.6
PyWavelets==1.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
segmentation-models==1.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0rc0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@b0b0e9d0cf74a7f7dc84ab4ace616121dc5acc1a
tensorflow-macos==2.13.0rc0
termcolor==2.3.0
tifffile==2023.4.12
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.6.0
urllib3==1.26.15
uvicorn==0.17.6
Werkzeug==2.3.4
wrapt==1.14.1
